parent 358af5c6001dda30d591259839082c864e16a385 author Yossi Farjoun <farjoun@broadinstitute.org> 1617479686 -0400 committer Yossi Farjoun <farjoun@broadinstitute.org> 1618455120 -0400
Change the way the script is run. instead of modfying the ENTRYPOINT, we will simply add the "shell" (default sh) and the "script_prefix" (default -c)
and then the script.

That way the possible comple entry point is not disturbed. if a user would like to remove the ENTRYPOINT they can provide an "option" "--entrypoint ''" for example.